Change `http://pypi...` to `https://pypi....`
I find when doing a wget on this file I get the Pypi link as `http://pypi.python.org` however this fails to install everytime when changed to `https://pypi.python.org` it installs fine. I'm not sure if this is something my end or the url needs to be https? Perhaps https is better in any case?